wip: kernel-netlink: Offload bypass policies for IKE ports on interfaces
It's not necessary to explicitly enable port_bypass, the regular socket
policies work fine to bypass the software policies.

wip: Once we tried to offload the bypass policies, we'd also know if
interfaces actually support packet/policy offloading so we could possibly
adapt the code when installing SAs/policies.  But might not simplify the
code much due to the locking required.